requirements.txt
# Tello Drone Photography System Dependencies

# Core SDK
djitellopy>=2.5

# Image Processing & Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# QR Code Detection
pyzbar>=0.1.9

# Pose/Gesture Detection (for emergency hand signal)
mediapipe>=0.10.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
Pillow>=10.0.0
